:hammer: 코드 컨벤션으로 변경
