+++ src/hooks/myPage/useDeleteUserAccount.ts
@@ -1,6 +1,14 @@
-import { deleteUserAccount } from '../../apis/myPage/myPageApi';
-import { DeleteUserAccountType, UseDeleteUserAccount } from '../../types';
+import {
+  deleteUserAccount,
+  DeleteUserAccountType,
+} from '../../apis/myPage/myPageApi';
+import { APIError, BookDiscussionInfo } from '../../types';
 import useMutate from '../useMutate';
+
+export interface UseDeleteUserAccount {
+  onSuccess?: (data: BookDiscussionInfo) => void;
+  onError?: (error: Error | APIError) => void;
+}
 
 function useDeleteUserAccount({ onSuccess, onError }: UseDeleteUserAccount) {
   const { mutate, data, error, isLoading } = useMutate<
