+++ src/hooks/myPage/useGetMyPageInfo.ts
@@ -1,4 +1,4 @@
-import getMypageInfo, { getMypageInfoProps } from '../../apis/myPage/myPageApi';
+import getMypageInfo, { getMypageInfoType } from '../../apis/myPage/myPageApi';
 import useQuery from '../useQuery';
 import { MyPageInfoProps, UseMyPageInfoDetailProps } from '../../types';
 
@@ -9,7 +9,7 @@ function useGetMyPageInfo({
   isErrorBoundary = false,
 }: UseMyPageInfoDetailProps) {
   const { data, isLoading, error, setData } = useQuery<
-    getMypageInfoProps,
+    getMypageInfoType,
     MyPageInfoProps
   >({
     fetchFn: getMypageInfo,
