+++ src/hooks/myPage/useUpdateUsername.ts
@@ -1,6 +1,15 @@
-import { updateUsername } from '../../apis/myPage/myPageApi';
-import { UpdateUsernameType, UseUpdateUsername } from '../../types';
+import {
+  updateUsername,
+  UpdateUsernameType,
+} from '../../apis/myPage/myPageApi';
+import { UserInfo } from '../../recoil/atoms';
+import { APIError } from '../../types';
 import useMutate from '../useMutate';
+
+export interface UseUpdateUsername {
+  onSuccess?: (data: UserInfo) => void;
+  onError?: (error: Error | APIError) => void;
+}
 
 function useUpdateUsername({ onSuccess, onError }: UseUpdateUsername) {
   const { mutate, data, error, isLoading } = useMutate<UpdateUsernameType, any>(
